backend/requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.12.1
boto3==1.28.34
botocore==1.31.34
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cli-helpers==2.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.3
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.13.2
djangorestframework==3.14.0
docutils==0.16
futures==3.0.5
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moto==4.2.2
packaging==23.1
pendulum==2.1.2
pgcli==3.5.0
pgspecial==2.1.0
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psycopg==3.1.10
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
rsa==4.9
s3==3.0.0
s3transfer==0.6.2
setproctitle==1.3.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tabulate==0.9.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
xmltodict==0.13.0
